+++ doc/changelog.rst
@@ -19,6 +19,7 @@ Features
 Bugs
 ----
 
+- Allowed IP address <-> VM association updates after manual VM configuration on hypervisor - `#168 <https://github.com/erigones/esdc-ce/issues/168>`__
 - Fixed problem with high amount of network traffic in the celeryev exchange - `#179 <https://github.com/erigones/esdc-ce/issues/179>`__
 - Fixed IP address validation, when multiple IPs are being added - `#177 <https://github.com/erigones/esdc-ce/issues/177>`__
 - Disable current compute in VM migration dialog - `#191 <https://github.com/erigones/esdc-ce/issues/191>`__
@@ -48,7 +49,6 @@ Bugs
 - Create required `domainmetadata` for every newly created domain - `#151 <https://github.com/erigones/esdc-ce/issues/151>`__
 - Do not display *pending* status when desired VM status was already reached - `#152 <https://github.com/erigones/esdc-ce/issues/152>`__
 - Fixed VM hostname fetching in `message_callback` (GUI/JS) - `#159 <https://github.com/erigones/esdc-ce/issues/159>`__
-- Allowed IP address <-> VM association updates after manual VM configuration on hypervisor - `#168 <https://github.com/erigones/esdc-ce/issues/168>`__
 
 
 2.5.2 (released on 2017-04-11)